Refactor finalfrontier-utils.
