+++ finalfrontier-utils/src/lib.rs
@@ -3,3 +3,9 @@ pub use crate::data::thread_data;
 
 mod progress;
 pub use crate::progress::FileProgress;
+
+mod util;
+pub use crate::util::{
+    common_config_from_matches, show_progress, skipgram_config_from_matches,
+    subword_config_from_matches, AppBuilder,
+};
